M7 Phase 5: construct-monomorphization layer + control-form/return boundary bridges
Lands the construct-monomorphization step of the carrier-bridge work, driving
the full suite green (1685/0). A #{Construct} constructor (some/none/ok/err) at
a plain call site whose own abi_bindings (or grounded call->type) resolve its
parametric result to a concrete by-value non-heap struct now builds the struct
directly (some__spec returning Option__int) instead of the carrier some() + a
carrier->concrete bridge at the by-value consumer. option-basic drops from 7
carrier crossings to 1.

To keep representations consistent end-to-end, this adds the boundary bridges a
by-value construct now needs when meeting the carrier ABI:
  - emit_fns.c: concrete->carrier RETURN bridge (main + emit_tail paths) via a
    heap malloc spill (a stack spill would dangle), for closure/thunk bodies
    whose C return is the uniform int64 carrier;
  - emit_expr.c emit_if_value: a by-value if-merge temp with per-arm
    carrier->concrete bridging (the if's own type is collapsed to the carrier);
  - expr_emits_byvalue_carrier_abi: control forms (if/do/let) delegate to the
    tail walker so a consumer of (if b (some 1) (none)) sees the by-value merge;
  - fn_body_tail_byvalue_carrier_type: recovers the concrete by-value type from
    the matched spec (the construct's e->type is carrier-collapsed);
  - find_matched_abi_spec: never apply the cross-spec fallback to a #{Construct}
    callee (mirrors emit_call_name), so a shared (some ...) Expr* is not reported
    by-value under a sibling int64-result spec;
  - the generic-carrier-param arg bridge fires when the callee param resolves to
    the int64 carrier C type (abstract (Option A)), not concrete by-value sinks.

Measured: raw suite-wide bridge-call count 1319 -> 1338 (neutral; the change
trades consumer-side carrier->concrete crossings for construct-side
concrete->carrier ones while reducing user construct sites). Full bridge
DELETION still requires monomorphizing the generic Option/Result CONSUMERS
(option-eq?/unwrap/...) too -- this lands the construct half and the reusable
boundary-bridge infrastructure the consumer half will build on.

+++ tests/fixtures/mutmap-typed-consumer/expected.c
@@ -4617,11 +4617,11 @@ static int64_t sum_hyvals(MutableMap__int__int * m, int64_t i, int64_t cap, int6
         if ((i) == (cap)) {
             return acc;
         } else {
-            if (mutmap_hyslot_hyoccupied_qu((int64_t)(intptr_t)(m), i)) {
+            if (mutmap_hyslot_hyoccupied_qu((int64_t)(intptr_t)((int64_t)(intptr_t)(m)), i)) {
                 MutableMap__int__int * __t52 = m;
                 int64_t __t53 = (i) + (INT64_C(1));
                 int64_t __t54 = cap;
-                int64_t __t55 = (acc) + (mutmap_hyslot_hyvalue((int64_t)(intptr_t)(m), i));
+                int64_t __t55 = (acc) + (mutmap_hyslot_hyvalue((int64_t)(intptr_t)((int64_t)(intptr_t)(m)), i));
                 m = __t52;
                 i = __t53;
                 cap = __t54;
@@ -4642,7 +4642,7 @@ static int64_t sum_hyvals(MutableMap__int__int * m, int64_t i, int64_t cap, int6
 }
 
 static int64_t total(MutableMap__int__int * m) {
-        return sum_hyvals(m, INT64_C(0), mutmap_hycap((int64_t)(intptr_t)(m)), INT64_C(0));
+        return sum_hyvals(m, INT64_C(0), mutmap_hycap((int64_t)(intptr_t)((int64_t)(intptr_t)(m))), INT64_C(0));
 }
 
 int main(int argc, char **argv) {
@@ -4658,12 +4658,12 @@ int main(int argc, char **argv) {
         {
             MutableMap__int__int * a_1014 = mutmap_new__spec__MutableMap__int__int__();
             (void)a_1014;
-            mutmap_hyset_ex((int64_t)(intptr_t)(a_1014), INT64_C(1), INT64_C(1), INT64_C(100));
-            mutmap_hyset_ex((int64_t)(intptr_t)(a_1014), INT64_C(2), INT64_C(2), INT64_C(200));
-            mutmap_hyset_ex((int64_t)(intptr_t)(a_1014), INT64_C(3), INT64_C(3), INT64_C(300));
+            mutmap_hyset_ex((int64_t)(intptr_t)((int64_t)(intptr_t)(a_1014)), INT64_C(1), INT64_C(1), INT64_C(100));
+            mutmap_hyset_ex((int64_t)(intptr_t)((int64_t)(intptr_t)(a_1014)), INT64_C(2), INT64_C(2), INT64_C(200));
+            mutmap_hyset_ex((int64_t)(intptr_t)((int64_t)(intptr_t)(a_1014)), INT64_C(3), INT64_C(3), INT64_C(300));
             printf("%lld\n", (long long)(total(a_1014)));
-            printf("%lld\n", (long long)(mutmap_hylen((int64_t)(intptr_t)(a_1014))));
-            mutmap_hyfree((int64_t)(intptr_t)(a_1014));
+            printf("%lld\n", (long long)(mutmap_hylen((int64_t)(intptr_t)((int64_t)(intptr_t)(a_1014)))));
+            mutmap_hyfree((int64_t)(intptr_t)((int64_t)(intptr_t)(a_1014)));
         }
         int64_t __t60;
         __t60 = INT64_C(0);
